feat: adicionado a página /home e a /login, e movido a antiga / para /scheduler, e implementado a rota privada em /scheduler
